renamed app to vmlog (was log)
