+++ vmlog/admin.py
