+++ vmlog/static/log/dialog.js
