+++ vmlog/test/test_decode.py
