+++ vmlog/test/test_serializer.py
